refactor(filesystem): extract fd-rooted primitives to local/fd_resolve.rs
local.rs was already 1005 lines pre-diff and had grown to ~1252 across
the fd-rooted traversal work — done last, after FIX 1-4, so those
changes moved with the extraction instead of needing a second pass.

The fd-primitive layer (ResolveError, open_one, dup_fd, resolve_walk,
descend_creating, remove_dir_all_fd/remove_dir_contents,
atomic_write_file, and helpers new_file_mode/new_dir_mode/
map_file_type/read_all/write_all) was already self-contained and
banner-commented as a distinct concern, with zero dependency on
DiskFilesystem/LocalMount/the RootFilesystem impl — every function
operates purely on BorrowedFd/OwnedFd/OsStr. Moved to
crates/ironclaw_filesystem/src/local/fd_resolve.rs (file + sibling
directory, no mod.rs needed). local.rs drops from ~1252 to 884 lines;
fd_resolve.rs is 451.

Visibility: extracted items the DiskFilesystem impl calls directly
(open_one, resolve_walk, descend_creating, remove_dir_all_fd,
atomic_write_file, map_file_type, new_file_mode, read_all, write_all,
resolve_error_to_filesystem_error) are pub(super); everything else
(dup_fd, remove_dir_contents, resolve_error_to_io, new_dir_mode,
MAX_REMOVE_DIR_DEPTH, LOCAL_WRITE_TEMP_COUNTER) stays private to the
submodule. io_error moved the other way — it's called from both
files but has more call sites in local.rs's RootFilesystem impl, so it
stays put and is now pub(super) for fd_resolve.rs to call back into.
stat_modified stays in local.rs: it's local-mount-stat-specific, not
part of the fd-primitive layer, and wasn't in the extraction list.

Re-ran FIX 2's architecture gate against the new shape: extended
GATED_FILES to scan both local.rs and local/fd_resolve.rs (the latter
has no #[cfg(test)] module of its own, so its entire contents count as
production), added a proof-of-binding test that plants a tokio::fs
import directly into fd_resolve.rs and confirms the gate still catches
it there — not just cosmetically listed. fd_resolve.rs's only
std::fs::* use (std::fs::File::from, wrapping an already-resolved fd)
was already on the allowlist.

Left the resolve_walk/descend_creating ancestor-walk duplication
(~10 shared lines: dup root, loop open_one over ancestors) alone —
tried a shared per-step-fallback helper and it read as more indirection
for less code, not more clarity, so kept the two functions as they are
per the "only if it genuinely reads better" guidance.

Verified: cargo test -p ironclaw_filesystem (32/32, including all 3
TOCTOU race tests) and -p ironclaw_architecture (all suites, including
the re-scoped gate, 8/8) pass; cargo clippy -p ironclaw_filesystem
-p ironclaw_architecture --all-features --tests is warning-free; Linux
cross-check (CC_x86_64_unknown_linux_gnu=x86_64-linux-gnu-gcc cargo
check -p ironclaw_filesystem --target x86_64-unknown-linux-gnu) is
clean; cargo fmt --check is clean.
